refactor: 進捗の計算の関数をコンポーネントから抽出 & グラフにラベルを追加
